# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Clinic API Documentation",
        "description": "API Documentation for Clinic management system",
        "contact": {
            "email": "support@clinic.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/slots": {
            "get": {
                "tags": [
                    "Slots"
                ],
                "summary": "Get list of slots",
                "description": "Returns list of slots",
                "operationId": "52b2e5f0f5a47bdeebd33aee49091640",
                "parameters": [
                    {
                        "name": "client_id",
                        "in": "query",
                        "description": "Client ID to filter slots",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "doctor_id",
                        "in": "query",
                        "description": "Doctor ID to filter slots",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Status to filter slots",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "available",
                                "booked",
                                "missed",
                                "done"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Slot"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Slots"
                ],
                "summary": "Create a new slot",
                "description": "Create a new slot",
                "operationId": "072807e7de7dcdb5aef91c18304bcb3a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Slot"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Slot created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Slot"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/slots/{id}": {
            "get": {
                "tags": [
                    "Slots"
                ],
                "summary": "Get slot by ID",
                "description": "Returns a single slot",
                "operationId": "f40ae8b7c5cf60f2aa364314df03f567",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of slot to return",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Slot"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Slot not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Slots"
                ],
                "summary": "Update an existing slot",
                "description": "Update an existing slot",
                "operationId": "21bc050040565f08041471cfd17fd7ec",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of slot to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Slot"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Slot updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Slot"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Slot not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Slots"
                ],
                "summary": "Delete a slot",
                "description": "Delete a slot",
                "operationId": "fbeb9ef289fcc9bf664c94b9a05e1ee2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of slot to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Slot deleted"
                    },
                    "404": {
                        "description": "Slot not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Slot": {
                "title": "Slot",
                "required": [
                    "doctor_id",
                    "start_time",
                    "end_time",
                    "status"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true
                    },
                    "doctor_id": {
                        "type": "integer"
                    },
                    "client_id": {
                        "type": "integer"
                    },
                    "start_time": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "end_time": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "available",
                            "booked",
                            "missed",
                            "done"
                        ]
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    }
                },
                "type": "object"
            }
        }
    },
...
}